requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
pydantic

llama-index==0.14.15
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-readers-file
pypdf
tiktoken